combined stress-common with common.js
